// Frontend/src/app/models/importInvoice.model.ts
export interface HoaDonNhap {
    maHDN: number;
    ngayNhap: string;
    trangThai: number;  
    maND: number;
    maNCC: number;

    hoTen: string;
    tenNCC: string;
    tongTien: number;
}
export interface CTHoaDonNhap {
    maCTHDN: number;
    soLuong: number;
    giaNhap: number;
    maSP: number;
    maHDN: number;

    tenSP: string;
    anhDaiDien: string;
}